Split module with scaling helpers
