+++ lib/src/snark/mod.rs
@@ -1,0 +1,4 @@
+
+mod snark;
+pub mod scaling_helpers;
+pub use snark::*;
